+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.7.0] - COMING
 
+### Added
+
+- Possibility to create pages to reuse any element you defined
+
 ### Changed
 
 - Patterns are in harmony (syntax respect some standards)